그로우룸 url 전반적 수정
그로우룸 조회
- Dto 생성
